Move and Rename files inside common/components
