+++ src/common/components/Footer/index.tsx
